% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_earthdata.R
\name{get_earthdata}
\alias{get_earthdata}
\title{Search NASA Earthdata Collections Metadata}
\usage{
get_earthdata(keyword, n_results, start_date = NULL, end_date = NULL)
}
\arguments{
\item{keyword}{Character. A search term or phrase used to find relevant datasets
in titles, descriptions, keywords, and provider names.}

\item{n_results}{Integer. The number of dataset entries to retrieve.
If more than 2000, multiple pages will be requested automatically.}

\item{start_date}{Character or NULL. Optional start date filter in "YYYY-MM-DD" format.
If provided, must be used with \code{end_date}.}

\item{end_date}{Character or NULL. Optional end date filter in "YYYY-MM-DD" format.
If provided, must be used with \code{start_date}.}
}
\value{
A data frame containing metadata about the matching datasets,
with only cleaned column names (columns with '.' or '$' removed).
}
\description{
Queries NASA's Common Metadata Repository (CMR) to search Earth science datasets
related to a specified keyword. Optionally filters results by a temporal range.
}
\details{
CMR is Earthdata’s search engine — the backend database that stores metadata about:
\itemize{
  \item Satellite datasets
  \item Earth science data (climate, ocean, atmosphere, land)
  \item Observational granules (single files like images, temperature readings, etc.)
  \item Services (subsetting, reformatting, and other data services)
}

The search finds matches based on the keyword provided. The keyword can appear in:
\itemize{
  \item Dataset titles
  \item Dataset descriptions
  \item Dataset keywords (tags)
  \item Some provider names
}


The function accesses the CMR API endpoint
\url{https://cmr.earthdata.nasa.gov/search/collections.json}.
It harmonizes columns across API pages and returns up to the number of requested results.
If no results are found, an empty data frame is returned.
}
\examples{
# Search for 10 datasets related to sea surface temperature
results <- get_earthdata(keyword = "sea surface temperature", n_results = 10)

# Search with a temporal constraint
results_time <- get_earthdata(
  keyword = "sea surface temperature",
  n_results = 10,
  start_date = "2020-01-01",
  end_date = "2020-01-02"
)

}
